[pig-seckill] 完成下单发送消息 创建订单

+++ pig-seckill/src/main/java/cn/com/k12code/pigseckill/goods/dto/BuyDTO.java
@@ -24,7 +24,7 @@ public class BuyDTO {
 	 * 商品数量
 	 */
 	@Min(value = 1)
-	private int itemCount;
+	private Long itemCount;
 
 	/**
 	 * 商品单价
